@@ -1,62 +1,69 @@
 {
   "name": "vega-editor",
-  "version": "0.46.1",
+  "version": "0.82.0",
   "homepage": "http://vega.github.io/editor",
   "private": true,
+  "license": "BSD-3-Clause",
   "devDependencies": {
-    "autoprefixer": "^9.6.1",
-    "awesome-typescript-loader": "^5.2.1",
-    "css-loader": "^3.2.0",
-    "gh-pages": "^2.1.1",
-    "html-webpack-plugin": "^3.2.0",
-    "husky": "^3.0.5",
-    "lint-staged": "^9.2.5",
-    "monaco-editor-webpack-plugin": "^1.7.0",
+    "@types/react-select": "^3.0.13",
+    "autoprefixer": "^9.8.2",
+    "css-loader": "^3.6.0",
+    "gh-pages": "^3.1.0",
+    "html-webpack-plugin": "^4.3.0",
+    "husky": "^4.2.5",
+    "lint-staged": "^10.2.11",
+    "monaco-editor-webpack-plugin": "^1.9.0",
     "postcss-loader": "^3.0.0",
-    "prettier": "^1.18.2",
-    "style-loader": "^1.0.0",
-    "typescript": "~3.6.3",
-    "vega-lite-dev-config": "^0.2.5",
-    "webpack": "4.40.2",
-    "webpack-bundle-analyzer": "^3.5.0",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1"
+    "style-loader": "^1.2.1",
+    "ts-loader": "^7.0.5",
+    "typescript": "~3.9.5",
+    "vega-lite-dev-config": "^0.11.7",
+    "webpack": "^4.43.0",
+    "webpack-bundle-analyzer": "^3.8.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
   },
   "dependencies": {
-    "@types/react": "^16.9.2",
-    "@types/react-redux": "^7.1.2",
-    "@types/react-router-dom": "^4.3.5",
-    "ajv": "^6.10.2",
-    "d3-geo-projection": "^2.7.0",
-    "history": "^4.10.1",
+    "@types/d3": "^5.7.2",
+    "@types/react": "^16.9.38",
+    "@types/react-redux": "^7.1.9",
+    "@types/react-resize-detector": "^4.2.0",
+    "@types/react-router-dom": "^5.1.5",
+    "@types/semver": "^7.2.0",
+    "ajv": "^6.12.2",
+    "d3-array": "^2.4.0",
+    "d3-geo-projection": "^2.9.0",
+    "d3-scale": "^3.2.1",
+    "file-loader": "^6.0.0",
+    "history": "^5.0.0",
     "json-stringify-pretty-compact": "^2.0.0",
     "lz-string": "^1.4.4",
-    "monaco-editor": "^0.18.1",
+    "monaco-editor": "^0.20.0",
     "prop-types": "^15.7.2",
-    "react": "^16.9.0",
+    "react": "^16.13.1",
     "react-clipboard.js": "^2.0.16",
-    "react-dom": "^16.9.0",
-    "react-feather": "^2.0.3",
-    "react-ga": "2.6.0",
-    "react-monaco-editor": "^0.31.0",
-    "react-paginate": "^6.3.0",
-    "react-portal": "^4.2.0",
-    "react-redux": "^7.1.1",
-    "react-router-dom": "^5.0.1",
-    "react-select": "^3.0.4",
-    "react-split-pane": "^0.1.87",
-    "react-tooltip": "^3.11.1",
-    "redux": "^4.0.4",
+    "react-dom": "^16.13.1",
+    "react-feather": "^2.0.8",
+    "react-monaco-editor": "^0.36.0",
+    "react-paginate": "^6.3.2",
+    "react-portal": "^4.2.1",
+    "react-redux": "^7.2.0",
+    "react-resize-detector": "^5.0.6",
+    "react-router-dom": "^5.2.0",
+    "react-select": "^3.1.0",
+    "react-split-pane": "^0.1.91",
+    "react-tooltip": "^4.2.7",
+    "redux": "^4.0.5",
     "redux-localstorage": "^0.4.1",
     "redux-thunk": "^2.3.0",
-    "tslib": "^1.10.0",
-    "vega-cli": "5.6.0",
-    "vega-datasets": "^1.25.0",
-    "vega-embed": "^5.1.3",
-    "vega-lite": "4.0.0-beta.2",
+    "tslib": "^2.0.0",
+    "vega-cli": "5.13.0",
+    "vega-datasets": "^1.31.1",
+    "vega-embed": "^6.9.0",
+    "vega-lite": "4.13.1",
     "vega-schema-url-parser": "^1.1.0",
-    "vega-themes": "^2.4.0",
-    "vega-tooltip": "^0.19.1"
+    "vega-themes": "^2.8.3",
+    "vega-tooltip": "^0.23.0"
   },
   "husky": {
     "hooks": {
