@@ -1,14 +1,18 @@
 import * as Monaco from 'monaco-editor/esm/vs/editor/editor.api';
 import * as React from 'react';
 import MonacoEditor from 'react-monaco-editor';
+import ReactResizeDetector from 'react-resize-detector';
+import {RouteComponentProps, withRouter} from 'react-router-dom';
 import {debounce} from 'vega';
 import {mapDispatchToProps, mapStateToProps} from '.';
-import {LAYOUT, Mode, SIDEPANE} from '../../constants';
+import {SIDEPANE} from '../../constants';
 import './config-editor.css';
 
-type Props = ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDispatchToProps>;
+type Props = ReturnType<typeof mapStateToProps> &
+  ReturnType<typeof mapDispatchToProps> &
+  RouteComponentProps<{compressed: string}>;
 
-export default class ConfigEditor extends React.PureComponent<Props> {
+class ConfigEditor extends React.PureComponent<Props> {
   public editor: Monaco.editor.IStandaloneCodeEditor;
   public handleEditorChange = (spec: string) => {
     const newSpec = spec === '' ? '{}' : spec;
@@ -25,8 +29,12 @@ export default class ConfigEditor extends React.PureComponent<Props> {
     if (!confirmation) {
       return;
     }
+    if (this.props.history.location.pathname !== '/edited') {
+      this.props.history.push('/edited');
+    }
     this.props.mergeConfigSpec();
   }
+
   public handleExtractConfig() {
     const confirmation = confirm('The spec and config will be formatted.');
     if (!confirmation) {
@@ -47,7 +55,7 @@ export default class ConfigEditor extends React.PureComponent<Props> {
       contextMenuOrder: 0,
       id: 'MERGE_CONFIG',
       label: 'Merge Config Into Spec',
-      run: this.handleMergeConfig.bind(this)
+      run: this.handleMergeConfig.bind(this),
     });
 
     editor.addAction({
@@ -55,11 +63,14 @@ export default class ConfigEditor extends React.PureComponent<Props> {
       contextMenuOrder: 1,
       id: 'EXTRACT_CONFIG',
       label: 'Extract Config From Spec',
-      run: this.handleExtractConfig.bind(this)
+      run: this.handleExtractConfig.bind(this),
     });
+
     this.editor = editor;
+
     if (this.props.sidePaneItem === SIDEPANE.Config) {
       this.editor.focus();
+      this.editor.layout();
     }
   }
 
@@ -69,50 +80,43 @@ export default class ConfigEditor extends React.PureComponent<Props> {
     }
   }
 
-  public componentWillReceiveProps(nextProps) {
-    if (nextProps.sidePaneItem === SIDEPANE.Config) {
+  public componentDidUpdate(prevProps, prevState) {
+    if (this.props.sidePaneItem === SIDEPANE.Config) {
       this.editor.focus();
+      this.editor.layout();
       this.props.setEditorReference(this.editor);
     }
   }
 
-  public getEditorHeight() {
-    // height of header : 60
-    // height of compiled Spec Header :30
-    let height = window.innerHeight - 60 - LAYOUT.MinPaneSize - 30; // 60 is the height of header;
-    if (this.props.compiledVegaSpec) {
-      height -= this.props.compiledVegaPaneSize - 30;
-    }
-    return height;
-  }
   public render() {
     return (
-      <div
-        className={this.props.mode === Mode.Vega ? 'full-height-wrapper' : ''}
-        style={{
-          display: this.props.sidePaneItem === SIDEPANE.Editor ? 'none' : ''
-        }}
-      >
+      <>
+        <ReactResizeDetector
+          handleWidth
+          handleHeight
+          onResize={(width: number, height: number) => {
+            this.editor.layout({width, height: height});
+          }}
+        ></ReactResizeDetector>
         <MonacoEditor
-          height={this.getEditorHeight()}
           options={{
             autoClosingBrackets: 'never',
             autoClosingQuotes: 'never',
-            automaticLayout: true,
             cursorBlinking: 'smooth',
             folding: true,
             lineNumbersMinChars: 4,
             minimap: {enabled: false},
             scrollBeyondLastLine: false,
-            wordWrap: 'on'
+            wordWrap: 'on',
           }}
-          ref="ConfigEditor"
           language="json"
           onChange={debounce(700, this.handleEditorChange)}
           value={this.props.configEditorString}
-          editorDidMount={e => this.handleEditorMount(e)}
+          editorDidMount={(e) => this.handleEditorMount(e)}
         />
-      </div>
+      </>
     );
   }
 }
+
+export default withRouter(ConfigEditor);
