@@ -3,18 +3,19 @@ import LZString from 'lz-string';
 import * as Monaco from 'monaco-editor/esm/vs/editor/editor.api';
 import * as React from 'react';
 import MonacoEditor from 'react-monaco-editor';
+import ReactResizeDetector from 'react-resize-detector';
 import {RouteComponentProps, withRouter} from 'react-router-dom';
 import {debounce} from 'vega';
 import parser from 'vega-schema-url-parser';
 import {mapDispatchToProps, mapStateToProps} from '.';
-import {EDITOR_FOCUS, KEYCODES, LAYOUT, Mode, SCHEMA, SIDEPANE} from '../../../constants';
+import {EDITOR_FOCUS, KEYCODES, Mode, SCHEMA, SIDEPANE} from '../../../constants';
 import './index.css';
 
 type Props = ReturnType<typeof mapStateToProps> &
   ReturnType<typeof mapDispatchToProps> &
   RouteComponentProps<{compressed: string}>;
 
-class Editor extends React.PureComponent<Props, {}> {
+class Editor extends React.PureComponent<Props> {
   public editor: Monaco.editor.IStandaloneCodeEditor;
   constructor(props: Props) {
     super(props);
@@ -44,8 +45,12 @@ class Editor extends React.PureComponent<Props, {}> {
     if (!confirmation) {
       return;
     }
+    if (this.props.history.location.pathname !== '/edited') {
+      this.props.history.push('/edited');
+    }
     this.props.mergeConfigSpec();
   }
+
   public handleExtractConfig() {
     const confirmation = confirm('The spec and config will be formatted.');
     if (!confirmation) {
@@ -72,7 +77,7 @@ class Editor extends React.PureComponent<Props, {}> {
     if (spec.$schema === undefined) {
       spec = {
         $schema: SCHEMA[Mode.Vega],
-        ...spec
+        ...spec,
       };
       if (confirm('Adding schema URL will format the specification too.')) {
         this.props.updateVegaSpec(stringify(spec));
@@ -85,7 +90,7 @@ class Editor extends React.PureComponent<Props, {}> {
     if (spec.$schema === undefined) {
       spec = {
         $schema: SCHEMA[Mode.VegaLite],
-        ...spec
+        ...spec,
       };
       if (confirm('Adding schema URL will format the specification too.')) {
         this.props.updateVegaLiteSpec(stringify(spec));
@@ -99,12 +104,13 @@ class Editor extends React.PureComponent<Props, {}> {
       editor.deltaDecorations(this.props.decorations, []);
       this.props.setEditorFocus(EDITOR_FOCUS.SpecEditor);
     });
+
     editor.addAction({
       contextMenuGroupId: 'vega',
       contextMenuOrder: 0,
       id: 'ADD_VEGA_SCHEMA',
       label: 'Add Vega schema URL',
-      run: this.addVegaSchemaURL.bind(this)
+      run: this.addVegaSchemaURL.bind(this),
     });
 
     editor.addAction({
@@ -112,7 +118,7 @@ class Editor extends React.PureComponent<Props, {}> {
       contextMenuOrder: 1,
       id: 'ADD_VEGA_LITE_SCHEMA',
       label: 'Add Vega-Lite schema URL',
-      run: this.addVegaLiteSchemaURL.bind(this)
+      run: this.addVegaLiteSchemaURL.bind(this),
     });
 
     editor.addAction({
@@ -120,7 +126,7 @@ class Editor extends React.PureComponent<Props, {}> {
       contextMenuOrder: 2,
       id: 'CLEAR_EDITOR',
       label: 'Clear Spec',
-      run: this.onClear.bind(this)
+      run: this.onClear.bind(this),
     });
 
     editor.addAction({
@@ -128,7 +134,7 @@ class Editor extends React.PureComponent<Props, {}> {
       contextMenuOrder: 3,
       id: 'MERGE_CONFIG',
       label: 'Merge Config Into Spec',
-      run: this.handleMergeConfig.bind(this)
+      run: this.handleMergeConfig.bind(this),
     });
 
     editor.addAction({
@@ -136,7 +142,7 @@ class Editor extends React.PureComponent<Props, {}> {
       contextMenuOrder: 4,
       id: 'EXTRACT_CONFIG',
       label: 'Extract Config From Spec',
-      run: this.handleExtractConfig.bind(this)
+      run: this.handleExtractConfig.bind(this),
     });
 
     editor.getModel().getOptions();
@@ -145,6 +151,7 @@ class Editor extends React.PureComponent<Props, {}> {
 
     if (this.props.sidePaneItem === SIDEPANE.Editor) {
       editor.focus();
+      editor.layout();
       this.props.setEditorFocus(EDITOR_FOCUS.SpecEditor);
     }
   }
@@ -160,8 +167,15 @@ class Editor extends React.PureComponent<Props, {}> {
   public editorWillMount(monaco: typeof Monaco) {
     const compressed = this.props.match.params.compressed;
     if (compressed) {
-      const spec = LZString.decompressFromEncodedURIComponent(compressed);
+      let spec: string = LZString.decompressFromEncodedURIComponent(compressed);
+
       if (spec) {
+        const newlines = (spec.match(/\n/g) || '').length + 1;
+        if (newlines <= 1) {
+          console.log('Formatting spec string from URL that did not contain newlines.');
+          spec = stringify(JSON.parse(spec));
+        }
+
         this.updateSpec(spec);
       } else {
         this.props.logError(new Error(`Failed to decompress URL. Expected a specification, but received ${spec}`));
@@ -169,24 +183,26 @@ class Editor extends React.PureComponent<Props, {}> {
     }
   }
 
-  public componentWillReceiveProps(nextProps: Props) {
-    if (nextProps.sidePaneItem === SIDEPANE.Editor) {
-      if (this.props.sidePaneItem !== nextProps.sidePaneItem) {
+  public componentDidUpdate(prevProps, prevState) {
+    if (this.props.sidePaneItem === SIDEPANE.Editor) {
+      if (prevProps.sidePaneItem !== this.props.sidePaneItem) {
         this.editor.focus();
-        this.props.setEditorReference(this.editor);
+        this.editor.layout();
+        prevProps.setEditorReference(this.editor);
       }
     }
 
-    if (this.props.view !== nextProps.view) {
-      this.props.compiledEditorRef && this.props.compiledEditorRef.deltaDecorations(this.props.decorations, []);
-      this.props.editorRef && this.props.editorRef.deltaDecorations(this.props.decorations, []);
+    if (prevProps.view !== this.props.view) {
+      prevProps.compiledEditorRef && prevProps.compiledEditorRef.deltaDecorations(prevProps.decorations, []);
+      prevProps.editorRef && prevProps.editorRef.deltaDecorations(prevProps.decorations, []);
     }
 
-    if (nextProps.parse) {
+    if (this.props.parse) {
       this.editor.focus();
-      this.updateSpec(nextProps.value);
-      this.props.setConfig(nextProps.configEditorString);
-      this.props.parseSpec(false);
+      this.editor.layout();
+      this.updateSpec(this.props.value);
+      prevProps.setConfig(this.props.configEditorString);
+      prevProps.parseSpec(false);
     }
   }
 
@@ -234,45 +250,34 @@ class Editor extends React.PureComponent<Props, {}> {
     }
   }
 
-  public getEditorHeight() {
-    // height of header : 60
-    // height of compiled Spec Header :30
-    let height = window.innerHeight - 60 - LAYOUT.MinPaneSize - 30; // 60 is the height of header;
-    if (this.props.compiledVegaSpec) {
-      height -= this.props.compiledVegaPaneSize - 30;
-    }
-    return height;
-  }
-
   public render() {
     return (
-      <div
-        className={this.props.mode === Mode.Vega ? 'full-height-wrapper' : ''}
-        style={{
-          display: this.props.sidePaneItem === SIDEPANE.Editor ? '' : 'none'
-        }}
-      >
+      <>
+        <ReactResizeDetector
+          handleWidth
+          handleHeight
+          onResize={(width: number, height: number) => {
+            this.editor.layout({width, height: height});
+          }}
+        ></ReactResizeDetector>
         <MonacoEditor
-          height={this.getEditorHeight()}
-          ref="editor"
           language="json"
           options={{
             autoClosingBrackets: 'never',
             autoClosingQuotes: 'never',
-            automaticLayout: true,
             cursorBlinking: 'smooth',
             folding: true,
             lineNumbersMinChars: 4,
             minimap: {enabled: false},
             scrollBeyondLastLine: false,
-            wordWrap: 'on'
+            wordWrap: 'on',
           }}
           value={this.props.value}
           onChange={debounce(700, this.handleEditorChange)}
           editorWillMount={this.editorWillMount}
           editorDidMount={this.editorDidMount}
         />
-      </div>
+      </>
     );
   }
 }
