@@ -20,33 +20,22 @@ export function mapStateToProps(state: State) {
     selectedExample: state.selectedExample,
     sidePaneItem: state.sidePaneItem,
     themeName: state.themeName,
-    value: state.editorString
+    value: state.editorString,
   };
 }
 
 export function mapDispatchToProps(dispatch: Dispatch<EditorActions.Action>) {
   return bindActionCreators(
     {
-      clearConfig: EditorActions.clearConfig,
       extractConfig: EditorActions.extractConfigSpec,
-      logError: EditorActions.logError,
       mergeConfigSpec: EditorActions.mergeConfigSpec,
-      parseSpec: EditorActions.parseSpec,
       setConfig: EditorActions.setConfig,
       setConfigEditorString: EditorActions.setConfigEditorString,
       setEditorReference: EditorActions.setEditorReference,
-      setSidePaneItem: EditorActions.setSidePaneItem,
       setThemeName: EditorActions.setThemeName,
-      toggleCompiledVegaSpec: EditorActions.toggleCompiledVegaSpec,
-      updateEditorString: EditorActions.updateEditorString,
-      updateVegaLiteSpec: EditorActions.updateVegaLiteSpec,
-      updateVegaSpec: EditorActions.updateVegaSpec
     },
     dispatch
   );
 }
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Renderer);
+export default connect(mapStateToProps, mapDispatchToProps)(Renderer);
