@@ -4,11 +4,13 @@ import Renderer from './renderer';
 
 export function mapStateToProps(state: State) {
   return {
+    baseURL: state.baseURL,
     config: state.config,
+    editorString: state.editorString,
     mode: state.mode,
     vegaLiteSpec: state.vegaLiteSpec,
     vegaSpec: state.vegaSpec,
-    view: state.view
+    view: state.view,
   };
 }
 
