@@ -1,6 +1,4 @@
 import {connect} from 'react-redux';
-import {bindActionCreators, Dispatch} from 'redux';
-import * as EditorActions from '../../../actions/editor';
 import {State} from '../../../constants/default-state';
 import Renderer from './renderer';
 
@@ -8,20 +6,9 @@ export function mapStateToProps(state: State) {
   return {
     handle: state.handle,
     isAuthenticated: state.isAuthenticated,
-    mode: state.mode
+    mode: state.mode,
+    private: state.private,
   };
 }
 
-export function mapDispatchToProps(dispatch: Dispatch<EditorActions.Action>) {
-  return bindActionCreators(
-    {
-      receiveCurrentUser: EditorActions.receiveCurrentUser
-    },
-    dispatch
-  );
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Renderer);
+export default connect(mapStateToProps)(Renderer);
