@@ -7,6 +7,7 @@ import * as EditorActions from '../../../actions/editor';
 import {EDITOR_FOCUS, LAYOUT} from '../../../constants';
 import {State} from '../../../constants/default-state';
 import CompiledSpecDisplayHeader from '../compiled-spec-header';
+import ReactResizeDetector from 'react-resize-detector';
 
 type Props = ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDispatchToProps>;
 
@@ -17,32 +18,29 @@ class CompiledSpecDisplay extends React.PureComponent<Props> {
     this.props.setCompiledEditorReference(this.editor);
   }
 
-  public componentDidUpdate(prevProps) {
-    if (this.props.compiledVegaPaneSize !== prevProps.compiledVegaPaneSize) {
-      if (this.editor) {
-        this.editor.layout();
-      }
-    }
-  }
-
   public render() {
     return (
-      <div className={'sizeFixEditorParent full-height-wrapper'}>
+      <div className={'full-height-wrapper'}>
         <CompiledSpecDisplayHeader />
+        <ReactResizeDetector
+          handleWidth
+          handleHeight
+          onResize={(width: number, height: number) => {
+            this.editor.layout({width, height: height - LAYOUT.MinPaneSize});
+          }}
+        ></ReactResizeDetector>
         <MonacoEditor
           height={this.props.compiledVegaPaneSize - LAYOUT.MinPaneSize}
           options={{
-            automaticLayout: true,
             folding: true,
             minimap: {enabled: false},
             readOnly: true,
             scrollBeyondLastLine: false,
-            wordWrap: 'on'
+            wordWrap: 'on',
           }}
-          ref="compiledEditor"
           language="json"
           value={stringify(this.props.value)}
-          editorDidMount={editor => {
+          editorDidMount={(editor) => {
             editor.onDidFocusEditorText(() => {
               this.props.compiledEditorRef && this.props.compiledEditorRef.deltaDecorations(this.props.decorations, []);
               this.props.editorRef && this.props.editorRef.deltaDecorations(this.props.decorations, []);
@@ -64,7 +62,7 @@ function mapStateToProps(state: State) {
     editorRef: state.editorRef,
     mode: state.mode,
     sidePaneItem: state.sidePaneItem,
-    value: state.vegaSpec
+    value: state.vegaSpec,
   };
 }
 
@@ -72,13 +70,10 @@ export function mapDispatchToProps(dispatch: Dispatch<EditorActions.Action>) {
   return bindActionCreators(
     {
       setCompiledEditorReference: EditorActions.setCompiledEditorRef,
-      setEditorFocus: EditorActions.setEditorFocus
+      setEditorFocus: EditorActions.setEditorFocus,
     },
     dispatch
   );
 }
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(CompiledSpecDisplay);
+export default connect(mapStateToProps, mapDispatchToProps)(CompiledSpecDisplay);
