@@ -1,12 +1,25 @@
 import {connect} from 'react-redux';
-
+import {bindActionCreators, Dispatch} from 'redux';
+import * as EditorActions from '../../actions/editor';
 import {State} from '../../constants/default-state';
 import Renderer from './renderer';
 
 export function mapStateToProps(state: State) {
   return {
-    view: state.view
+    signals: state.signals,
+    view: state.view,
   };
 }
 
-export default connect(mapStateToProps)(Renderer);
+export function mapDispatchToProps(dispatch: Dispatch<EditorActions.Action>) {
+  return bindActionCreators(
+    {
+      addSignal: EditorActions.addSignal,
+      setSignals: EditorActions.setSignals,
+      setView: EditorActions.setView,
+    },
+    dispatch
+  );
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Renderer);
