@@ -7,6 +7,13 @@ import {formatValueLong} from '../table/renderer';
 interface Props {
   view: View;
   signal: string;
+  onValueChange: (key, value) => void;
+  maskListener: boolean;
+  isHovered: boolean;
+  isTimelineSelected: boolean;
+  clickedSignal: any;
+  hoverValue: any;
+  timeline: boolean;
   onClickHandler?: (header: string) => void;
 }
 
@@ -15,10 +22,11 @@ interface State {
 }
 
 export default class SignalRow extends React.PureComponent<Props, State> {
+  private listenerAttached = false;
   constructor(props) {
     super(props);
     this.state = {
-      signalValue: this.props.view.signal(this.props.signal)
+      signalValue: this.props.view.signal(this.props.signal),
     };
     this.signalHandler = this.signalHandler.bind(this);
   }
@@ -26,22 +34,68 @@ export default class SignalRow extends React.PureComponent<Props, State> {
     if (prevProps.view !== this.props.view) {
       prevProps.view.removeSignalListener(prevProps.signal, this.signalHandler);
       this.props.view.addSignalListener(this.props.signal, this.signalHandler);
-      this.setState({
-        signalValue: this.props.view.signal(this.props.signal)
-      });
+      this.setState(
+        {
+          signalValue: this.props.view.signal(this.props.signal),
+        },
+        () => this.props.onValueChange(this.props.signal, this.props.view.signal(this.props.signal))
+      );
     }
   }
   public componentDidMount() {
-    this.props.view.addSignalListener(this.props.signal, this.signalHandler);
+    if (!this.props.maskListener) {
+      this.props.view.addSignalListener(this.props.signal, this.signalHandler);
+      this.listenerAttached = true;
+    }
   }
   public componentWillUnmount() {
     this.props.view.removeSignalListener(this.props.signal, this.signalHandler);
+    this.listenerAttached = false;
   }
+
+  public componentWillReceiveProps(nextProps) {
+    if (nextProps.maskListener && this.listenerAttached) {
+      this.props.view.removeSignalListener(this.props.signal, this.signalHandler);
+      this.listenerAttached = false;
+    } else if (!this.listenerAttached && !nextProps.maskListener) {
+      this.props.view.addSignalListener(this.props.signal, this.signalHandler);
+      this.listenerAttached = true;
+    }
+  }
+
+  public renderSignal = () => {
+    const {isTimelineSelected, isHovered, clickedSignal, hoverValue} = this.props;
+    if (isTimelineSelected && isHovered) {
+      return hoverValue;
+    }
+    if (isTimelineSelected) {
+      return clickedSignal;
+    } else if (isHovered) {
+      return hoverValue;
+    } else {
+      return null;
+    }
+  };
+
+  public getBackgroundColor = () => {
+    if (this.props.isTimelineSelected && this.props.isHovered) {
+      return '#fce57e';
+    }
+    if (this.props.isTimelineSelected && this.props.clickedSignal !== undefined) {
+      return '#A4F9C8';
+    } else if (this.props.isHovered && this.props.hoverValue !== undefined) {
+      return '#fce57e';
+    } else {
+      return '';
+    }
+  };
+
   public render() {
     let tooLong = false;
     let formatted = '';
+    const value = this.renderSignal();
     if (!isDate(this.state.signalValue)) {
-      const formatValue = formatValueLong(this.state.signalValue);
+      const formatValue = formatValueLong(value ? value : this.state.signalValue);
       if (formatValue !== undefined) {
         tooLong = formatValue.tooLong;
         formatted = formatValue.formatted;
@@ -51,16 +105,24 @@ export default class SignalRow extends React.PureComponent<Props, State> {
       }
     } else {
       tooLong = false;
-      formatted = new Date(this.state.signalValue).toUTCString();
+      formatted = new Date(value ? value : this.state.signalValue).toUTCString();
     }
     if (tooLong) {
       return (
         <tr>
-          <td onClick={() => this.props.onClickHandler && this.props.onClickHandler(this.props.signal)}>
+          <td
+            className="pointer"
+            onClick={() => this.props.onClickHandler && this.props.onClickHandler(this.props.signal)}
+          >
             {this.props.signal}
             <Search />
           </td>
-          <td title="The field is too large to be displayed. Please use the view API (see JS console).">
+          {this.props.timeline && <td style={{padding: 0}}>{this.props.children}</td>}
+          <td
+            style={{backgroundColor: this.getBackgroundColor()}}
+            key={this.props.signal}
+            title="The field is too large to be displayed. Please use the view API (see JS console)."
+          >
             <span>(...)</span>
           </td>
         </tr>
@@ -68,19 +130,37 @@ export default class SignalRow extends React.PureComponent<Props, State> {
     } else {
       return (
         <tr>
-          <td onClick={() => this.props.onClickHandler && this.props.onClickHandler(this.props.signal)}>
+          <td
+            style={{whiteSpace: 'nowrap'}}
+            className="pointer"
+            onClick={() => this.props.onClickHandler && this.props.onClickHandler(this.props.signal)}
+          >
             {this.props.signal}
             <Search />
           </td>
-          <td key={this.props.signal}>{formatted}</td>
+          {this.props.timeline && <td style={{padding: 0}}>{this.props.children}</td>}
+          <td
+            style={{
+              whiteSpace: 'nowrap',
+              backgroundColor: this.getBackgroundColor(),
+            }}
+            key={this.props.signal}
+          >
+            {formatted}
+          </td>
         </tr>
       );
     }
   }
 
   private signalHandler(signalName: string, currentValue) {
-    this.setState({
-      signalValue: currentValue
-    });
+    this.setState(
+      {
+        signalValue: currentValue,
+      },
+      () => {
+        this.props.onValueChange(this.props.signal, currentValue);
+      }
+    );
   }
 }
