@@ -1,6 +1,6 @@
 import * as d3 from 'd3-geo-projection';
 
-export default function(vegaProjections: any) {
+export default function (vegaProjections: any) {
   for (const p of [
     'airy',
     'aitoff',
@@ -28,6 +28,7 @@ export default function(vegaProjections: any) {
     'eisenlohr',
     'fahey',
     'foucaut',
+    'foucautSinusoidal',
     'gilbert',
     'gingery',
     'ginzburg4',
@@ -42,6 +43,15 @@ export default function(vegaProjections: any) {
     'healpix',
     'hill',
     'homolosine',
+    'hufnagel',
+    'hyperElliptical',
+    'interruptedBoggs',
+    'interruptedHomolosine',
+    'interruptedMollweide',
+    'interruptedMollweideHemispheres',
+    'interruptedSinuMollweide',
+    'interruptedSinusoidal',
+    'interruptedQuarticAuthalic',
     'kavrayskiy7',
     'lagrange',
     'larrivee',
@@ -54,20 +64,25 @@ export default function(vegaProjections: any) {
     'modifiedStereographicGs50',
     'modifiedStereographicMiller',
     'modifiedStereographicLee',
-    'mollweide',
+    // 'mollweide', // included in vega-projection
     'mtFlatPolarParabolic',
     'mtFlatPolarQuartic',
     'mtFlatPolarSinusoidal',
-    'naturalEarth1',
     'naturalEarth2',
     'nellHammer',
+    'nicolosi',
     'patterson',
     'polyconic',
+    'polyhedralButterfly',
+    'polyhedralCollignon',
+    'polyhedralWaterman',
+    'gringortenQuincuncial',
+    'peirceQuincuncial',
     'rectangularPolyconic',
     'robinson',
     'satellite',
-    'sinusoidal',
     'sinuMollweide',
+    'sinusoidal',
     'times',
     'twoPointAzimuthalUsa',
     'twoPointEquidistantUsa',
@@ -75,22 +90,12 @@ export default function(vegaProjections: any) {
     'vanDerGrinten2',
     'vanDerGrinten3',
     'vanDerGrinten4',
+    'wagner',
     'wagner4',
     'wagner6',
     'wagner7',
     'wiechel',
     'winkel3',
-    'interruptedHomolosine',
-    'interruptedSinusoidal',
-    'interruptedBoggs',
-    'interruptedSinuMollweide',
-    'interruptedMollweide',
-    'interruptedMollweideHemispheres',
-    'polyhedralButterfly',
-    'polyhedralCollignon',
-    'polyhedralWaterman',
-    'gringortenQuincuncial',
-    'peirceQuincuncial'
   ]) {
     vegaProjections(p, d3['geo' + p[0].toUpperCase() + p.slice(1)]);
   }
