@@ -2,28 +2,28 @@ const shortcuts = [
   {
     mac: '⌘ + b',
     text: 'Parse the specification and render visualization',
-    windows: 'Ctrl + b'
+    windows: 'Ctrl + b',
   },
   {
     mac: '⌘ + ?',
     text: 'Open the help window',
-    windows: 'Ctrl + ?'
+    windows: 'Ctrl + ?',
   },
   {
     mac: '⌘ + Space',
     text: 'Open Intellisense in the editor',
-    windows: 'Ctrl + Space'
+    windows: 'Ctrl + Space',
   },
   {
     mac: '⌘ + f11',
     text: 'Toggle fullscreen mode',
-    windows: 'Ctrl + f11'
+    windows: 'Ctrl + f11',
   },
   {
     mac: '⌘ + Shift + o',
     text: `Navigate to a symbol in the specification`,
-    windows: 'Ctrl + Shift + o'
-  }
+    windows: 'Ctrl + Shift + o',
+  },
 ];
 
 export default shortcuts;
